+++ classes/image_converter.hpp
@@ -1,16 +1,18 @@
-// fconvert v2.1.0 (c) 2023 - 2026 Eraldo Bako - MIT License
+// fconvert v2.2.0 (c) 2023 - 2026 Eraldo Bako - MIT License
 // Maintaier: eraldobako@gmail.com
-#ifndef IMAGE_CONVERTER_H
-#define IMAGE_CONVERTER_H
+#ifndef IMAGE_CONVERTER_HPP
+#define IMAGE_CONVERTER_HPP
 
 #include <filesystem>
 #include <string>
+
+#include <opencv2/core.hpp> // #include <opencv2/opencv.hpp>
+
 #ifdef _WIN32
     #include <libraw.h>
 #else
     #include <libraw/libraw.h>
 #endif
-#include <opencv2/opencv.hpp>
 
 cv::Mat read_camera_raw(const std::string& raw_path);
 void image_convert_logic(std::filesystem::path in, std::string fmt, bool silent);