+++ classes/video_converter.hpp
@@ -1,7 +1,7 @@
-// fconvert v2.1.0 (c) 2023 - 2026 Eraldo Bako - MIT License
+// fconvert v2.2.0 (c) 2023 - 2026 Eraldo Bako - MIT License
 // Maintaier: eraldobako@gmail.com
-#ifndef VIDEO_CONVERTER_H
-#define VIDEO_CONVERTER_H
+#ifndef VIDEO_CONVERTER_HPP
+#define VIDEO_CONVERTER_HPP
 
 #include <filesystem>
 #include <string>